test(chat): split media send coverage within size limits
